fix: rename package imports from zync_app to nunakin_app
Replace all package:zync_app/ references with package:nunakin_app/
across lib/, integration_test/, test/ and scripts/.
Build verified: app-debug.apk compiles successfully.
