+++ lib/core/error/exceptions.dart
