+++ lib/core/models/user_status.dart
