+++ lib/core/network/network_info.dart
