+++ lib/features/auth/domain/repositories/auth_repository.dart
