+++ scripts/fix_firebase_emojis.dart
@@ -6,8 +6,8 @@
 
 import 'package:firebase_core/firebase_core.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:zync_app/firebase_options.dart';
-import 'package:zync_app/core/models/user_status.dart';
+import 'package:nunakin_app/firebase_options.dart';
+import 'package:nunakin_app/core/models/user_status.dart';
 
 Future<void> main() async {
   print('🔧 Iniciando reparación de emojis en Firebase...\n');
